+++ src/main/java/com/bulletphysics/collision/broadphase/DispatcherInfo.java
@@ -7,11 +7,11 @@
  * This software is provided 'as-is', without any express or implied warranty.
  * In no event will the authors be held liable for any damages arising from
  * the use of this software.
- * 
- * Permission is granted to anyone to use this software for any purpose, 
+ *
+ * Permission is granted to anyone to use this software for any purpose,
  * including commercial applications, and to alter it and redistribute it
  * freely, subject to the following restrictions:
- * 
+ *
  * 1. The origin of this software must not be misrepresented; you must not
  *    claim that you wrote the original software. If you use this software
  *    in a product, an acknowledgment in the product documentation would be
@@ -27,26 +27,26 @@ import com.bulletphysics.linearmath.IDebugDraw;
 
 /**
  * Current state of {@link Dispatcher}.
- * 
+ *
  * @author jezek2
  */
 public class DispatcherInfo {
 
-	public float timeStep;
+	public double timeStep;
 	public int stepCount;
 	public DispatchFunc dispatchFunc;
-	public float timeOfImpact;
+	public double timeOfImpact;
 	public boolean useContinuous;
 	public IDebugDraw debugDraw;
 	public boolean enableSatConvex;
 	public boolean enableSPU = true;
 	public boolean useEpa = true;
-	public float allowedCcdPenetration = 0.04f;
+	public double allowedCcdPenetration = 0.04f;
 	//btStackAlloc*	m_stackAllocator;
 
 	public DispatcherInfo() {
 		dispatchFunc = DispatchFunc.DISPATCH_DISCRETE;
 		timeOfImpact = 1f;
 	}
-	
+
 }
